feat(QTM-723): Typography converted from styled components to css modules
GlobalStyle was rename to Typography
