+++ .storybook/preview.js
@@ -1,11 +1,12 @@
-import GlobalStyle from '../components/GlobalStyle';
+import Typography from '../components/Typography';
 import '../components/theme.css';
 import '../components/_utilities.css';
+import '../components/global.css';
 
 const CSSDecorator = (Story, context) => (
   <>
     <Story {...context} />
-    <GlobalStyle />
+    <Typography />
   </>
 );
