--- components/GlobalStyle/GlobalStyle.jsx
@@ -1,5 +1,0 @@
-import Typography from './sub-components/Typography';
-
-const GlobalStyle = () => <Typography />;
-
-export default GlobalStyle;
