--- components/GlobalStyle/GlobalStyle.unit.test.jsx
@@ -1,21 +1,0 @@
-import { render, screen } from '@testing-library/react';
-
-import GlobalStyle from './GlobalStyle';
-
-describe.skip('Global Style', () => {
-  it('should apply global style on h1 element', () => {
-    render(
-      <>
-        <GlobalStyle />
-        <h1>Hello!</h1>
-      </>,
-    );
-
-    expect(screen.getByRole('heading')).toHaveStyle({
-      margin: 0,
-      'font-weight': 'normal',
-      'font-size': '36px',
-      'letter-spacing': '0.25px',
-    });
-  });
-});
