--- components/GlobalStyle/index.d.ts
@@ -1,7 +1,0 @@
-import { FC } from 'react';
-
-declare const GlobalStyle: FC;
-declare const Typography: FC;
-
-export { Typography };
-export default GlobalStyle;
