--- components/GlobalStyle/index.js
@@ -1,8 +1,0 @@
-import GlobalStyle from './GlobalStyle';
-import Typography from './sub-components/Typography';
-import Title from './sub-components/Title';
-import Subtitle from './sub-components/Subtitle';
-import Blockquote from './sub-components/Blockquote';
-
-export { Typography, Title, Subtitle, Blockquote };
-export default GlobalStyle;
