--- components/GlobalStyle/sub-components/Blockquote.jsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const Blockquote = styled.blockquote`
-  font-size: 12px;
-  line-height: 18px;
-  letter-spacing: 0.2px;
-  margin: 0;
-`;
-
-export default Blockquote;
