--- components/GlobalStyle/sub-components/Subtitle.jsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const Subtitle = styled.span`
-  font-size: 14px;
-  line-height: 21px;
-  letter-spacing: 0.3px;
-`;
-
-export default Subtitle;
