--- components/GlobalStyle/sub-components/Title.jsx
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-const Title = styled.h1``;
-
-export default Title;
