+++ components/Typography/index.css
@@ -1,0 +1,2 @@
+@import "./sub-components/Blockquote.module.css";
+@import "./sub-components/Subtitle.module.css";
