+++ components/Typography/index.d.ts
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+
+declare const Typography: FC;
+
+export default Typography;
