+++ components/Typography/index.js
@@ -1,0 +1,7 @@
+import Typography from './Typography';
+import Title from './sub-components/Title';
+import Subtitle from './sub-components/Subtitle';
+import Blockquote from './sub-components/Blockquote';
+
+export { Title, Subtitle, Blockquote };
+export default Typography;
