+++ components/Typography/sub-components/Blockquote.module.css
@@ -1,0 +1,6 @@
+.blockquote {
+  font-size: 12px;
+  line-height: 18px;
+  letter-spacing: 0.2px;
+  margin: 0;
+}
