+++ components/Typography/sub-components/Subtitle.module.css
@@ -1,0 +1,5 @@
+.span {
+  font-size: 14px;
+  line-height: 21px;
+  letter-spacing: 0.3px;
+}
